services/requirements-dev.txt
# Union of all service dependencies — used by CI lint/test job
fastapi>=0.100,<1
uvicorn>=0.23,<1
httpx>=0.24,<1
pydantic>=2.0,<3
boto3>=1.28,<2
tiktoken>=0.5,<1
opentelemetry-api>=1.20,<2
opentelemetry-sdk>=1.20,<2
opentelemetry-exporter-otlp-proto-grpc>=1.20,<2
opentelemetry-instrumentation-fastapi>=0.41b0,<1
opentelemetry-instrumentation-httpx>=0.41b0,<1
opentelemetry-instrumentation-botocore>=0.41b0,<1
